+++ R/utils.R
@@ -201,6 +201,10 @@ separate_markers <- function(
 ) {
   # data.bk <- data
   # sep <- "__"
+  # generate.ref.alt <- TRUE
+  # biallelic= TRUE
+
+
 
   # check if markers column is present
   if (!tibble::has_name(data, "MARKERS")) {
@@ -1316,8 +1320,12 @@ gt_recoding <- function(x, gt = TRUE, gt.bin = TRUE, gt.vcf = TRUE, gt.vcf.nuc =
           A1 = dplyr::recode(REF, "A" = "001", "C" = "002", "G" = "003", "T" = "004"),
           A2 = dplyr::recode(ALT, "A" = "001", "C" = "002", "G" = "003", "T" = "004")
         )
-      remove.extra <- TRUE
       }
+
+      if (all(!rlang::has_name(x, "A1"), !rlang::has_name(x, "REF"))) {
+        x  %<>% dplyr::mutate(A1 =  "001", A2 = "002")
+      }
+      remove.extra <- TRUE
     }